+++ Assets/LDtkUnity/Runtime/Data/Interfaces/ILDtkUid.cs
@@ -2,6 +2,6 @@
 {
     internal interface ILDtkUid
     {
-        long Uid { get; }
+        int Uid { get; }
     }
 }
